requirements.txt
transformers
torch
numpy<2.0
pillow
albumentations
pycocotools
scikit-learn
accelerate
torchvision
torchmetrics
git+https://github.com/LLaVA-VL/LLaVA-NeXT.git
einops
wandb
hydra-core
matplotlib
tqdm
pytest
pytest-cov
pytest-mock
nltk

#flash-attn
#python 3.12